Change sorting of course blocks
When I was looking through the sort function in rodatraden_blocks.js I
noticed that the function comment didn't match the function body. It
said that it sorted in ascending order based on the start of the course,
but the code
sorted based in descending order based on the length of the course.

I liked both versions so I combined them. This is a matter of personal
preference but I like that the first courses in the scedule are the
earliest courses, and if they start at the same time I want the longest
one on top. Having the longest on top makes it easier to fit other
courses below it as well.

Since the courses were sorted in place I also removed the return
statement to make that more clear.