Replace matrix with a 90 degrees rotated grid
I didn't like the varaible name "matrix". To me that is a matematical
matrix with a fixed size. Therefore I renamed it to "grid". All of the
functions has also gotten new
names and updated JSDoc.

The orientation also didn't align with the visual representation,
meaning that the first index into the 2D structure wasn't the row index.
I've swapped that around now.

A few of the loops did double duty of also checking array boudaries.
I've moved that into separate if statements that also prints console
errors.

I've kept a "TEMP" comment to remind that a course parameter needs to be
renamed outside the file as well.