Modify render function to use a Map
The render function has been edited to clarify what it
actually does. It does two things: group courses by year and determines
the vertical position of each course.

I've changed code to use a Map instead of using an array as a
dictionary. This makes the intention clearer. I've also moved the Map
creation into a separete function.

I can't return the Map as it is without changing code outside the file
so for now I convert the Map into the old structure that callers of the
function expects. I also can't rename the function yet for the same
reason.