Decrease use of absolute positions and fix button
Now only the actual course blocks rely on absolute positions.

Also fixed the add year button that I managed to break. I added the
onclick eventlistener before adding the button. I also had to move that
button out of the div so that it didn't mix with the rows of each year.