Refactory all D3.js logic
This is a large commit. I've moved all D3 code that updates DOM elements
into separate functions. In doing so I have moved some code around,
refactored it and changed som things. It's a lot to sum up so I'll let
the code and comments speak for itself.

A noteworthy change is the academic year footer. I've changed how it
looks when as user isn't logged in, which also applies to block
schedules that the user doesn't own. Before the buttons still shown but
without text which made them look clickable. The hover effect also still
applied. Now the buttons have been replaced by a wide bar of similar
look that spans the entire academic year.