Fix issues adding years
There was a bug where you couldn't add a year to an empty block
schedule. Now the function to add years uses the starting year of the
block schedule as a fallback to be able to create the first year.

There was also a function that I forgot to rename everywhere. The result
was that the DOM elements didn't update when a new year was added.

I've also finally descided how to load the dataset variables defined in
the script tag. The comment at the top of the documents explains it.