Render block schedule per term
Now each term is a separate container. This opens up the possibility to
stack the terms vertically.

I've had to make changes to all D3 related functions to make it work.
I've not implemented the couse splitting yet so the schedule is still
rendered the same as before.

I've also had to change how the height of a block of courses is setup
as well as how the ects sum for each period is calculated and when.