Convert scripts to modules
Convert all script tags related to the block-schedule to modules. This
isolates the code better when only what is needed is exported. Global
can still be created but must be done so explicitly by
creating attributes on the window object. Also, modules are imported
asynchronously while the DOM is being
processed but  executed afterwards, like scripts with the defer tag.
It's also convenient to only define one script tag in the HTML and
import the remaining files from within the JavaScript.