Add Term class
I've not only added the Term class but also moved some functions around.

Also, now the heights of each course container is adapting to the
stacking of terms. When terms are side-by-side then their heights match.
When they are stacked then each term is only as high as it needs to be
to fit all courses in the term.

Also fixed a few bugs previously introduces when creating classes.