Replace year with pace in course-list modal
I've replaced year and length with weeks and pace. I've also appended
the ECTS to the title of each course.

Als ochanged from a HTML table to unordered lists. This gives us more
flexibility with the layout.

I haven't modified the table for private course-occasions. I'll do that
when I've finalized the design.