Show offset into period
There is now an extra section for each row in the course-list modal.
This is a list that for now only contains how many weeks into the period
the course starts, if it isn't zero. This list could be extended to show
if a course is only available every other year, for example.

Also made the ECTS span an inline-block so that the number and
its postfix stays together.