Position course-list filters responsively
The five filter selections were previously hardcoded to take up one
fifth of the width of the container. They dont have a width anymore.
Instead the container is a grid where the number of columns is decided
by a minimum width. This makes sure the selections aren't too small,
adding rows as needed to the grid.

I've also removed some Bootstrap boilerplate that made the filter
container overflow.

Finally, I added a mising placeholder text in one of the selections.