Replace table with custom solution
The auto-generated table has now been replaced with a responsive
solution. On a wide screen the layout is the same as before. As the
window get narrower, more and more columns disappear.

The information in
the removed columns are instead displayed near the name of the course.
The ECTS is displayed inline with the title and the small icons gets
listed horizontally below.

It is yet not possible to sort the courses as before. Instead all
courses are sorted by title in ascending order.