Re-implement sorting courses
The header now again contain links to change how the courses are sorted,
similar to how it worked before with the third-party table app.
JavaScript updates the links on pageload to
reflect how the courses are currently sorted.

I also changed the generic view to a custom one in preparation for
re-implementing pagination.