Reimplement course pagination
The pagination now been recreated. I've made some small changes though.
The next and previous buttons now only contain the symbol and not the
word, and there is a smal gap between them and the other link. This
separates highlights their differing functionality. Also, the ellipses
are no longer the same shape as the other buttons. They are more minimal
so as to not blend in with the numbered links.