Make pagination responsive
The pagination overflows slightly when the window is narrow. This is
solved by responsively hiding up to two links, depending on the which
page is active and how many there are in total. The responsiveness is
added on pageload by some JavaScript.

Also removes the global constants and improved documentation, as well
as adding bake the margin between the course list and pagination.