Switch link behavior in course adding modal
Clicking on the course title now adds the course. A tooltip shows up
when hovering to convince the user of this behaviour. The plus sign has
been replaced with a "read more" link, that takes the user to a page to
read more about the course-occasion. The same changes have been made to
the private section as well.

I've also added aria-labels to the links, and a "aria-hidden" attribute
to the ECTS after the course title since the ECTS is already present in
the aria-label.