Control font-size with container queries
The font-size is no longer tided to the window width. Instead we use
container queries that use the size of each course block individually to
switch between two font-sizes.

To make more space I've hidden the button to remove courses until the
user hovers over the course. They are always visible on mobile and
depending on the size of the course block, the text is either to the
right of the button or below it. I had to turn on word-wrap to break
longer words due to the larger font-size.

Also realized that xMax represented the number of weeks in a term. I've
renamed it to reflect that.

Also made sure that there is a gap between the terms when they are
stacked vertically, similar to the gap between academic years.