Add custom start week form field
The starting week of a private course is now chosen via two form fields,
startign period and weeks into the period start. This combines into
the starting week that gets stored. The starting year is also on the
same row as these new fields. This groups all fields that define the
start of the course.

I've also replaced all referenced to "length" with either "weeks" or
"length in weeks".

The note field also got a verbose name since it was missing one.