Move year and note field in form
I've changed the form for
private courses so that the academic year once again is a separate
field. I did that bacause I now show the title of the year, which is
more verbose and clear. The years now appear in ascending order.

I also moved the note field to the bottom since it isn't madatory.