Make small tweaks to private course-occasions
Re-ordered the columns to match the editing modal, and sorted the
private course occasions by title.

Also changed how the column for the start week looks like. If the start
week is 13, instead of displaying the number in the column we instead
show "LP2 - 3 veckor in". This is more descriptive and avoids having to
explain to the user how the start week works under the hood.

Finally, the note field now spans five rows, which emphasises that the
user is able to write a lot of text there.