Fix edge case when importing schedule
The year difference between the schedules was calculated from the first
course in the imported schedule. In the edge case that the imported
schedule starts with one or more empty years, the calculation is wrong.
Now the calculation uses the start_year member of the block for both the
created and the imported block.