Create BlockSchedule class
To render a block schedule you now interact with a BlockSchedule class.
This includes all relevant data and some public functions that you have
the option of using. This way the functions aren't spread out and at the
same time optional to use.

A small tweak to the BlockScheduleData is that it now updates itself on
initialization.