Generate schedule for profile detail page
Replaces the block schedule image with a scaled-down generated clickable
schedule. It doesn't support multiple schedules yet.

The base block courses are represented as a gray block and the first two
years are hidden. The term headers are also changed to not show specific
years.

Moved the margin and scale into the class so that it could be customized
for each schedule. This allowed me to scale down the schedule on the
profile detail without affecting anywhere else.