Implement sorting for private course occasions
The same sorting as the list of courses has now been implemented for the
list of private course occasions.

I moved the common sorting functions into a separate JavaScript so that
the different implementations of the custom table won't need to
duplicate code.

I also had to adjust the width of header and their media queries when
the caret also had to fit.