Add import button to block view
There is now a button in the menu that opens up a modal with a list of
block schedules. Selecting one of them imports all of its course
occasions.

The list of schedules is reused from the list of courses modal but the
class names have not yet been generalized to fit both uses.