Add new prerequisite model
This model will eventually replace the existing prerequisite model. They
will both exist until the data has been transfered from the old to the
new model.

Right now the prerequisites are only displayed in the course editing
modal for those courses that have the new type of prerequisite. It is
not possible to add or remove them yet.