Allow for the removal of prerequisites
Now it is possible to remove the new type of prerequisites. To do that
the fields are not bound to the form but instead only rendered in the
template.

I also removed the compress method of the prerequisite form since it
will never be used if the fields aren't bound to any form.