Show and migrate new prerequisites
A tool has been added that migrates all old prerequisites to the new
system. This tool, as well as the remnants of the old system of
prerequisites should be removed once the migration is complete.

The new prerequisites are also visualized on the course detail
page, alongside the old system. The courses a coruse gives access to is
also duplicated but using the new system, and it is now sorted
alphabetically.