Show prerequisite warning icon and tooltip.
The temporary background color has now been replaced with showing a
warning icon and a tooltip for each course with unmet prerequisites.

I've used exit and enter d3 selections instead of just appending the
icon. This is to be prepared for when the block schedule is updated
using AJAX instead of page refreshed.

Also adjusted the remove course button without affecting the look of it,
just the underlying structure.