Create Prerequisite JS class
To be able to list unmet prerequisites I changed the JSON representation
of the prerequisite and created a JS class to hold not only the
equivalent courses but the private key of the prerequisite.

With these changes to prerequisite logic could be rewritten to take
advantage of the new class.