Remove course block through AJAX
This is a large commit containing mostly changes for removing course
blocks using a GET request and not having to reload the whole page.

There are also many more changes that had to be made to facilitate this.
The need to re-render caused a lot of code to be moved around.

Also fixed a bug that hadn't been tested for where a function I had
changed in previous commits were used on the profile detail page.

Also fixed the indentation in urls.py and added type hints for the
request parameter in views.py.