Replace modalForm for course adding
The modalForm stops behaving after a number of course add/remove actions
and opening/closing the modal multiple times. After a while the
backdrop remains or the modal is reopened after having been closed.

The modal is now rendered directly usign a GET request. The animations
are still missing though.