Fix error when splitting course blocks
When splitting a course occasion into two blocks, the cloned block had
no height. Furthermore when widening the screen again the block did not
return to its original size. These issues are fixed now.

Also, the length has been renamed to visibleWeeks to signal its intended
use. It is not the length of the course occasion but the length of the
block representing the course occasion.