Add drag and drop functionality
Now course occasions can be moved using drag and drop. Ghost blocks appear as drop zones and a valid drop replaces the old course occasion with the new one.

The drag and drop code should be moved into a separate function though.

There is also starting to become a lot of duplication between BlockSchedule and BlockScheduleData.