Add block schedule view settings to profiles
The block schedules related to a profile can be very different. For some it is only neccessary to show the elective courses and for others all courses need to be shown. There is also the exception of the base block. Therefore the block schedule view of each profile needs to be different.

Now there is a setting per profile to set how the block schedule should be shown.